Vagrantfile: Add support for multiple machines
Create a cell of a group of AFS servers (3 by default) and one AFS
client. Variables at the top let you customize the number of servers
and cell/realm name.  Provisioning is improved so it only runs
ansible-playbook on the final server configured, so it only needs to
run once, but sets up the whole cell.  Used the following trick to do
that:
https://www.vagrantup.com/docs/provisioning/ansible#tips-and-tricks